CRM/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
httpx
pyotp
fpdf2>=2.7.8
PyYAML
psutil
cryptography
requests
